refactor(gpu): replace device-aware memcpy/is-device-ptr hooks with gpu_api functions (#508)
The runtime installed two global function-pointer hooks (g_device_aware_memcpy,
g_is_device_pointer) defined in clio_run_cxx (ipc_manager.cc) but written from
clio_run_cxx_gpu (gpu2cpu_init_*.cc). Since clio_run_cxx PRIVATE-links the GPU
lib, that back-reference formed a clio_run_cxx <-> clio_run_cxx_gpu cycle that
Linux's .so loader tolerates but Windows DLL linking cannot (LNK1120).

These are just the runtime face of GpuApi::Memcpy/IsDevicePointer, so make them
plain header functions in gpu_api.h instead of globals:
- ctp::DeviceAwareMemcpy() — host-fast-path + non-blocking-stream async copy
  (the old hook lambda), std::memcpy on non-GPU builds.
- ctp::IsDevicePointer() — thin wrapper over GpuApi::IsDevicePointer.

Drop the hook install code in gpu2cpu_init_{hip,sycl}.cc, delete the global
definitions in ipc_manager.cc, delete device_memcpy.h (+ chimaera shim), and
repoint bdev_runtime/worker/ipc_gpu2cpu to ctp::. No shared global -> no
cross-DLL back-edge -> the GPU library links standalone on Windows (verified:
clio_run_cxx_gpu.dll builds, 0 unresolved). Linux behavior unchanged.
