--- context-runtime/include/chimaera/device_memcpy.h
@@ -1,7 +1,0 @@
-// Backward-compat forwarding shim.
-// Prefer the new path: <clio_runtime/device_memcpy.h>.
-//
-// This header was relocated as part of the chimaera -> clio_runtime
-// rebrand. The original include path remains valid for now; the canonical
-// header lives at the new path.
-#include <clio_runtime/device_memcpy.h>
